♻️ Migrate the model about surveillance into init module of sub-package *config*.
